添加网络扫描功能

+++ src/renderer/views/Home.vue
@@ -9,21 +9,30 @@
             </a-button>
             <a-button icon="setting" size="small" style="float: left; margin-top: 5px;">系统配置</a-button>
             <a-button :icon="showParams ? 'up' : 'down' " size="small" style="float: right; margin-top: 5px;" @click="showParams=!showParams">启动参数</a-button>
-
         </div>
         <div class="more-control" v-if="showParams">
-            <a-divider orientation="left" class="btn-divider">启动参数</a-divider>
-            <a-tooltip title="码率越大越清晰，但是延迟也会增加。"><span class="btn-title">码率:</span></a-tooltip>
-            <a-radio-group buttonStyle="solid" size="small" v-model="execParams.bitRate">
-                <a-radio-button value="16M">16M</a-radio-button>
-                <a-radio-button value="8M">8M</a-radio-button>
-                <a-radio-button value="4M">4M</a-radio-button>
-                <a-radio-button value="2M">2M</a-radio-button>
-            </a-radio-group>
-            <a-tooltip title=""><span class="btn-title">最大帧数:</span></a-tooltip>
-            <a-input style="width: 50px;" size="small" :value="execParams.maxFps" />
-            <a-tooltip title="息屏：Ctrl+o；解锁：Ctrl+p"><span class="btn-title">允许息屏:</span></a-tooltip>
-            <a-switch :defaultChecked="execParams.turnScreenOff" size="small" @change="(checked) => {execParams.turnScreenOff=checked}" />
+            <div>
+                <a-divider orientation="left" class="btn-divider">启动参数</a-divider>
+                <a-tooltip title="码率越大越清晰，但是延迟也会增加。"><span class="btn-title-first">码率:</span></a-tooltip>
+                <a-radio-group buttonStyle="solid" size="small" v-model="execParams.bitRate">
+                    <a-radio-button value="16M">16M</a-radio-button>
+                    <a-radio-button value="12M">12M</a-radio-button>
+                    <a-radio-button value="8M">8M</a-radio-button>
+                    <a-radio-button value="4M">4M</a-radio-button>
+                    <a-radio-button value="2M">2M</a-radio-button>
+                </a-radio-group>
+                <a-tooltip title=""><span class="btn-title">最大帧数:</span></a-tooltip>
+                <a-input style="width: 50px;" size="small" :value="execParams.maxFps" />
+            </div>
+            <div>
+                <a-tooltip title="息屏：Ctrl+o；解锁：Ctrl+p"><span class="btn-title-first">允许息屏:</span></a-tooltip>
+                <a-switch :defaultChecked="execParams.turnScreenOff" size="small" class="btn-switch" @change="(checked) => {execParams.turnScreenOff=checked}" />
+                <a-tooltip title="息屏：Ctrl+o；解锁：Ctrl+p"><span class="btn-title">无线连接:</span></a-tooltip>
+                <a-switch :defaultChecked="wireless" size="small" class="btn-switch" @change="(checked) => {if(checked){scanWifi();} wireless=checked;}" />
+                <a-select style="width: 150px" placeholder="IP地址" showSearch :filterOption="true" size="small" v-if="wireless">
+                    <a-select-option v-for="ip in ipList" :key="ip" value="ip">{{ip}}</a-select-option>
+                </a-select>
+            </div>
         </div>
         <div class="main-log">
             <a-list class="log-list" :class="{'log-list-less': showParams}" size="small" bordered :dataSource="logList">
@@ -37,6 +46,8 @@
 
 <script>
     const { spawn } = require('child_process')
+    const os = require('os')
+    const net = require('net')
     const paramLabel = {
         bitRate: {
             param: '-b', type: 'string'
@@ -64,12 +75,14 @@
                     // maxFps: "15",
                     turnScreenOff: false,
                 },
+                wireless: false,
+                ipList: ["扫描中..."],
             }
         },
         watch: {
             workerProcess: {
                 handler(newVal, oldVal) {
-                    this.running = !newVal.killed;
+                    this.running = oldVal.killed == newVal.killed ? this.running : !newVal.killed;
                     // if(newVal.killed != oldVal.killed){
                     //     this.running = !newVal.killed;
                     // }
@@ -121,8 +134,11 @@
 
                 // 打印错误的后台可执行程序输出
                 this.workerProcess.stderr.on('data', function (data) {
+                    data = data.toString()
                     let ePos = data.indexOf("Exception");
-                    data = data.substring(0, ePos);
+                    if(ePos > -1) {
+                        data = data.substring(0, ePos);
+                    }
                     that.running = true;
                     that.logList.unshift({
                         data: data,
@@ -135,7 +151,7 @@
                 this.workerProcess.on('close', function (code) {
                     that.running = false;
                     that.logList.unshift({
-                        data: "退出 : " + code,
+                        data: "关闭 : " + code,
                         time: (new Date()).Format("yyyy-MM-dd hh:mm:ss"),
                         kind: "close"
                     });
@@ -151,7 +167,80 @@
                 })
             },
             closeExec: function () {
-                this.workerProcess.kill('SIGTERM');
+                if(this.workerProcess.killed){
+                    this.running = false;
+                }else {
+                    this.workerProcess.kill('SIGTERM');
+                }
+            },
+            scanWifi: function () {
+                let that = this;
+                let Socket = net.Socket
+                let port  = 5555
+                let scan = function(host, cb) {
+                    var socket = new Socket()
+                    var status = null
+                    socket.setTimeout(1500)
+                    socket.on('connect', function() {
+                        socket.end()
+                        cb && cb(null, host)
+                    })
+                    socket.on('timeout', function() {
+                        socket.destroy()
+                        cb && cb(new Error('timeout'), host)
+                    })
+                    socket.on('error', function(err) {
+                        cb && cb(err, host)
+                    })
+                    socket.on('close', function(err) {
+                    })
+                    socket.connect(port, host)
+                }
+                let cnt = 0;
+                let myipList = [];
+                //待扫描的开始网段，可换成192.168.0
+                let myIpList = this.getIpAddress()
+                for(let ipIndex = 0; ipIndex< myIpList.length; ipIndex++) {
+                    let myIp = myIpList[ipIndex]
+                    let ip = myIp.split(".").slice(0, 3).join(".")
+                    for (let i = 1; i < 255; i++) {
+                        scan(ip + '.' + i, function (err, host) {
+                            if (err) {
+                                that.logList.unshift({
+                                    data: host + ": " + err.message,
+                                    time: (new Date()).Format("yyyy-MM-dd hh:mm:ss"),
+                                    kind: "error"
+                                });
+                            } else {
+                                that.logList.unshift({
+                                    data: "找到ip: " + host,
+                                    time: (new Date()).Format("yyyy-MM-dd hh:mm:ss"),
+                                    kind: "success"
+                                });
+                                myipList.push(host);
+                            }
+                            cnt += 1;
+                            console.log('cnt: ' + cnt);
+                            if (cnt >= 254 * myIpList.length) {
+                                that.ipList = myipList;
+                            }
+                        })
+                    }
+                }
+            },
+            getIpAddress: function () {
+                let interfaces = os.networkInterfaces();
+                let ipList = []
+                for (var devName in interfaces) {
+                    var iface = interfaces[devName];
+                    for (var i = 0; i < iface.length; i++) {
+                        var alias = iface[i];
+                        if (alias.family === 'IPv4' && alias.address !== '127.0.0.1' && !alias.internal) {
+                            ipList.push(alias.address);
+                        }
+                    }
+                }
+                return ipList;
             }
         }
     }
@@ -180,26 +269,39 @@
     #home .more-control {
 
     }
+    #home .more-control > div {
+        padding: 5px 0px;
+    }
     #home .more-control .btn-title{
         margin-left: 10px;
-        font-size: 13px;
+        font-size: 14px;
+    }
+    #home .more-control .btn-switch {
+        margin-top: -4px;
     }
 
     #home .more-control .btn-divider {
         margin: 2px 0 10px 0;
-        font-size: 14px;
+        font-size: 13px;
     }
 
     #home .log-list {
         margin-top: 10px;
         height: 690px;
         overflow-y: auto;
+        font-size: 12px;
+    }
+
+    #home .log-list .ant-list-item {
+        border-bottom: 1px solid #F5F5F5;
+        padding: 3px 5px;
     }
     #home .log-list-less {
         height: 635px;
     }
 
-    #home .log-list .log-start {
+    #home .log-list .log-start,
+    #home .log-list .log-success {
         color: #42b983;
     }
